PROJECT/ai-financial-assistant/requirements.txt
streamlit
ibm-watson
transformers
torch
pandas
numpy
matplotlib
seaborn
scikit-learn
openai
granite-sdk
python-dotenv